Make it possible to nest lists and dictionaries multiple times
