+++ echo/__init__.py
@@ -1,4 +1,3 @@
 from .core import *  # noqa
-from .list import *  # noqa
+from .containers import *  # noqa
 from .selection import *  # noqa
-from .dict import *  # noqa
